# SA-OL/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ent.Eatinghistory:
    properties:
      added_time:
        description: AddedTime holds the value of the "added_time" field.
        type: string
      edges:
        $ref: '#/definitions/ent.EatinghistoryEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EatinghistoryQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.EatinghistoryEdges:
    properties:
      foodmenu:
        $ref: '#/definitions/ent.Foodmenu'
        description: Foodmenu holds the value of the foodmenu edge.
        type: object
      mealplan:
        $ref: '#/definitions/ent.Mealplan'
        description: Mealplan holds the value of the mealplan edge.
        type: object
      taste:
        $ref: '#/definitions/ent.Taste'
        description: Taste holds the value of the taste edge.
        type: object
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
        type: object
    type: object
  ent.Foodmenu:
    properties:
      addedtime:
        description: Addedtime holds the value of the "addedtime" field.
        type: string
      calories:
        description: Calories holds the value of the "calories" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.FoodmenuEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FoodmenuQuery when eager-loading is set.
        type: object
      groupid:
        description: Groupid holds the value of the "groupid" field.
        type: integer
      id:
        description: ID of the ent.
        type: integer
      menuing:
        description: Menuing holds the value of the "menuing" field.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
      typeid:
        description: Typeid holds the value of the "typeid" field.
        type: integer
    type: object
  ent.FoodmenuEdges:
    properties:
      eatinghistorys:
        description: Eatinghistorys holds the value of the eatinghistorys edge.
        items:
          $ref: '#/definitions/ent.Eatinghistory'
        type: array
      owner:
        $ref: '#/definitions/ent.User'
        description: Owner holds the value of the owner edge.
        type: object
    type: object
  ent.Mealplan:
    properties:
      date:
        description: Date holds the value of the "date" field.
        type: string
      edges:
        $ref: '#/definitions/ent.MealplanEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MealplanQuery when eager-loading is set.
        type: object
      food_id:
        description: FoodID holds the value of the "food_id" field.
        type: integer
      id:
        description: ID of the ent.
        type: integer
      meal_id:
        description: MealID holds the value of the "meal_id" field.
        type: integer
    type: object
  ent.MealplanEdges:
    properties:
      eatinghistorys:
        description: Eatinghistorys holds the value of the eatinghistorys edge.
        items:
          $ref: '#/definitions/ent.Eatinghistory'
        type: array
      owner:
        $ref: '#/definitions/ent.User'
        description: Owner holds the value of the owner edge.
        type: object
    type: object
  ent.Taste:
    properties:
      edges:
        $ref: '#/definitions/ent.TasteEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TasteQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      taste:
        description: Taste holds the value of the "taste" field.
        type: string
    type: object
  ent.TasteEdges:
    properties:
      eatinghistorys:
        description: Eatinghistorys holds the value of the eatinghistorys edge.
        items:
          $ref: '#/definitions/ent.Eatinghistory'
        type: array
    type: object
  ent.User:
    properties:
      edges:
        $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      eatinghistorys:
        description: Eatinghistorys holds the value of the eatinghistorys edge.
        items:
          $ref: '#/definitions/ent.Eatinghistory'
        type: array
      foodmenus:
        description: Foodmenus holds the value of the foodmenus edge.
        items:
          $ref: '#/definitions/ent.Foodmenu'
        type: array
      mealplans:
        description: Mealplans holds the value of the mealplans edge.
        items:
          $ref: '#/definitions/ent.Mealplan'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API Eatinghistory
  version: "1.0"
paths:
...
  /tastes:
    get:
      description: list taste entities
      operationId: list-taste
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Taste'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List taste entities
    post:
      consumes:
      - application/json
      description: Create taste
      operationId: create-taste
      parameters:
      - description: Taste entity
        in: body
        name: taste
        required: true
        schema:
          $ref: '#/definitions/ent.Taste'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Taste'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create taste
  /tastes/{id}:
    delete:
      description: get taste by ID
      operationId: delete-taste
      parameters:
      - description: Taste ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a taste entity by ID
    get:
      description: get taste by ID
      operationId: get-taste
      parameters:
      - description: Taste ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Taste'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a taste entity by ID
    put:
      consumes:
      - application/json
      description: update taste by ID
      operationId: update-taste
      parameters:
      - description: Taste ID
        in: path
        name: id
        required: true
        type: integer
      - description: Taste entity
        in: body
        name: taste
        required: true
        schema:
          $ref: '#/definitions/ent.Taste'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Taste'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a taste entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"